src/common/requirements.txt
feedgenerator
requests
beautifulsoup4
Pillow
aws_lambda_powertools[tracer]